requirements.txt
Pillow==8.2.0
click==7.1.2
